reworked day 16 to use a helper class
